Implementando métodos do MatriculadoDAO.

+++ src/main/java/br/ufg/inf/fabrica/conporta022018/persistencia/GenericoDAO.java
@@ -17,7 +17,7 @@ public abstract class GenericoDAO<T> {
 
     }
 
-    public T buscar(Long id) {
+    public T buscar(int id) {
         return null;
     }
